requirements.txt
cycler==0.10.0
kiwisolver==1.2.0
matplotlib==3.2.1
numpy==1.18.4
Pillow==7.2.0
pyparsing==2.4.7
python-dateutil==2.8.1
six==1.14.0
